@@ -6,11 +6,17 @@
     "start": "ng serve --open",
     "build": "ng build"
   },
-  "dependencies": {
-    "@angular/core": "^16.0.0"
-  },
   "devDependencies": {
     "@angular-devkit/build-angular": "^20.3.9",
-    "typescript": "~5.0.0"
+    "@angular/cli": "^20.3.9",
+    "@types/jasmine": "^5.1.12",
+    "typescript": "5.8"
+  },
+  "dependencies": {
+    "@angular/common": "^20.3.10",
+    "@angular/core": "^20.3.10",
+    "@angular/forms": "^20.3.10",
+    "@angular/platform-browser": "^20.3.10",
+    "@angular/platform-browser-dynamic": "^20.3.10"
   }
 }
