@@ -1,13 +1,10 @@
 import { Component, Input } from '@angular/core';
+import { CommonModule } from '@angular/common';
 
 @Component({
   selector: 'app-menu-list',
-  template: `
-    <h2>Menu Items</h2>
-    <ul>
-      <li *ngFor="let item of items">{{ item }}</li>
-    </ul>
-  `
+  templateUrl: './menu-list.component.html',
+  imports: [CommonModule],
 })
 export class MenuListComponent {
   @Input() items: string[] = ['Sample Item 1', 'Sample Item 2'];
